refactor: migrate schemas to src dir
