+++ features/authorization/EmailSignupForm.tsx
@@ -2,7 +2,7 @@ import { useForm } from 'react-hook-form';
 import { Alert, Button } from '@mui/material';
 import { PasswordInput } from '../../components/textfield/PasswordInput';
 import { useRouter } from 'next/router';
-import { SignupPostResolver } from '../../schemas/SignupSchema';
+import { SignupPostResolver } from '@/schemas/SignupSchema';
 import { mutate } from 'swr';
 import { useState } from 'react';
 import { BrothTextField } from '@/components/textfield/BrothTextField';
